requirements.txt
django==3.0.6
djangorestframework==3.11.0
psycopg2==2.8.5
gunicorn==20.0.4
django-heroku==0.3.1